requirements.txt
streamlit
langchain
openai
chromadb
pypdf
tiktoken
python-dotenv
